+++ mkdocs/markdowns/changelog.md
@@ -33,5 +33,5 @@
 ## [0.0.15] - 2021-9-18
 
 - add go,kegg annotion to rnaseq-basic analysis
-- add pipeline denove rna-seq analysis
+- add pipeline denovo rna-seq analysis
 - add pipeline rnaseq lncRNA analysis